Move integration tests out of calamari, and integration tests support for manta
