+++ runtime/integration-tests/Cargo.toml
@@ -1,0 +1,78 @@
+[package]
+authors = ["Manta Network"]
+edition = "2021"
+homepage = "https://manta.network"
+license = "GPL-3.0"
+name = "integration-tests"
+repository = "https://github.com/Manta-Network/Manta/"
+version = "4.0.5"
+
+[dependencies]
+calamari-runtime = { path = "../calamari", optional = true }
+manta-runtime = { path = "../manta", optional = true }
+
+[dev-dependencies]
+codec = { package = "parity-scale-codec", version = '3.1.2', default-features = false, features = ["derive", "max-encoded-len"] }
+scale-info = { version = "2.1.2", default-features = false, features = ["derive"] }
+lazy_static = "1.4.0"
+frame-system = { git = 'https://github.com/paritytech/substrate.git', branch = "polkadot-v0.9.28" }
+frame-support = { git = 'https://github.com/paritytech/substrate.git', branch = "polkadot-v0.9.28" }
+pallet-balances = { git = 'https://github.com/paritytech/substrate.git', branch = "polkadot-v0.9.28" }
+pallet-democracy = { git = 'https://github.com/paritytech/substrate.git', branch = "polkadot-v0.9.28" }
+pallet-collective = { git = 'https://github.com/paritytech/substrate.git', branch = "polkadot-v0.9.28" }
+pallet-session = { git = 'https://github.com/paritytech/substrate.git', branch = "polkadot-v0.9.28" }
+pallet-transaction-payment = { git = 'https://github.com/paritytech/substrate.git', branch = "polkadot-v0.9.28" }
+pallet-scheduler = { git = 'https://github.com/paritytech/substrate.git', branch = "polkadot-v0.9.28" }
+pallet-assets = { git = "https://github.com/paritytech/substrate.git", default-features = false, branch = "polkadot-v0.9.28" }
+pallet-utility = { git = "https://github.com/paritytech/substrate.git", default-features = false, branch = "polkadot-v0.9.28" }
+pallet-xcm = { git = 'https://github.com/paritytech/polkadot.git', branch = "release-v0.9.28" }
+sp-arithmetic = { git = 'https://github.com/paritytech/substrate.git', default-features = false, branch = "polkadot-v0.9.28" }
+sp-core = { git = 'https://github.com/paritytech/substrate.git', default-features = false, branch = "polkadot-v0.9.28" }
+sp-runtime = { git = 'https://github.com/paritytech/substrate.git', default-features = false, branch = "polkadot-v0.9.28" }
+sp-io = { git = 'https://github.com/paritytech/substrate.git', default-features = false, branch = "polkadot-v0.9.28" }
+sp-std = { git = 'https://github.com/paritytech/substrate.git', default-features = false, branch = "polkadot-v0.9.28" }
+xcm = { git = 'https://github.com/paritytech/polkadot.git', branch = "release-v0.9.28" }
+xcm-builder = { git = 'https://github.com/paritytech/polkadot.git', branch = "release-v0.9.28" }
+xcm-executor = { git = 'https://github.com/paritytech/polkadot.git', branch = "release-v0.9.28" }
+
+polkadot-core-primitives = { git = 'https://github.com/paritytech/polkadot.git', branch = "release-v0.9.28" }
+polkadot-parachain = { git = 'https://github.com/paritytech/polkadot.git', branch = "release-v0.9.28" }
+polkadot-runtime-parachains = { git = 'https://github.com/paritytech/polkadot.git', branch = "release-v0.9.28" }
+xcm-simulator = { git = 'https://github.com/paritytech/polkadot.git', branch = "release-v0.9.28" }
+
+cumulus-pallet-dmp-queue = { git = 'https://github.com/paritytech/cumulus.git', branch = "polkadot-v0.9.28" }
+cumulus-pallet-parachain-system = { git = 'https://github.com/paritytech/cumulus.git', branch = "polkadot-v0.9.28" }
+cumulus-pallet-xcm = { git = 'https://github.com/paritytech/cumulus.git', branch = "polkadot-v0.9.28" }
+cumulus-pallet-xcmp-queue = { git = 'https://github.com/paritytech/cumulus.git', branch = "polkadot-v0.9.28" }
+cumulus-primitives-parachain-inherent = { git = "https://github.com/paritytech/cumulus.git", branch = "polkadot-v0.9.28" }
+cumulus-test-relay-sproof-builder = { git = "https://github.com/paritytech/cumulus.git", branch = "polkadot-v0.9.28" }
+parachain-info = { git = 'https://github.com/paritytech/cumulus.git', branch = "polkadot-v0.9.28" }
+
+# Orml dependencies
+orml-traits = { git = "https://github.com/manta-network/open-runtime-module-library.git", branch = "polkadot-v0.9.28" }
+orml-xtokens = { git = "https://github.com/manta-network/open-runtime-module-library.git", branch = "polkadot-v0.9.28" }
+
+# Self dependencies
+pallet-asset-manager = { path = '../../pallets/asset-manager' }
+runtime-common = { path = '../common', features = ["test-helpers"] }
+
+calamari-vesting = { path = '../../pallets/vesting', default-features = false }
+manta-collator-selection = { path = '../../pallets/collator-selection', default-features = false }
+manta-primitives = { path = '../../primitives/manta', default-features = false }
+manta-support = { package = "pallet-manta-support", path = "../../pallets/manta-support", default-features = false }
+# Self dependencies
+pallet-manta-pay = { path = '../../pallets/manta-pay', default-features = false, features = ["runtime"] }
+pallet-manta-sbt = { path = '../../pallets/manta-sbt', default-features = false, features = ["runtime"] }
+pallet-parachain-staking = { path = '../../pallets/parachain-staking', default-features = false }
+pallet-tx-pause = { path = '../../pallets/tx-pause', default-features = false }
+session-key-primitives = { path = '../../primitives/session-keys', default-features = false }
+nimbus-primitives = { git = "https://github.com/manta-network/nimbus.git", tag = "v4.0.1", default-features = false }
+
+[features]
+default = ["calamari"]
+calamari = [
+    "calamari-runtime"
+]
+manta = [
+    "manta-runtime"
+]
